requirements.txt
requests==2.32.3
python-dotenv==1.0.1
ffmpeg-python==0.2.0
